Introduce genesis block that uses RevVault.
