+++ casper/src/main/scala/coop/rchain/casper/genesis/Genesis.scala
@@ -138,7 +138,7 @@ object Genesis {
     )
 
   private def withContracts[F[_]: Concurrent](
-      blessedTerms: List[DeployData],
+      blessedTerms: Seq[DeployData],
       initial: BlockMessage,
       startHash: StateHash,
       runtimeManager: RuntimeManager[F]
