+++ casper/src/test/scala/coop/rchain/casper/genesis/contracts/TestUtil.scala
@@ -1,17 +1,22 @@
 package coop.rchain.casper.genesis.contracts
 
-import cats.{FlatMap, Parallel}
 import cats.effect.{Concurrent, ContextShift, Sync}
+import cats.implicits._
+import cats.{FlatMap, Parallel}
+import coop.rchain.casper.MultiParentCasperTestUtil.createBonds
+import coop.rchain.casper.genesis.Genesis
 import coop.rchain.casper.protocol.{BlockMessage, DeployData}
 import coop.rchain.casper.util.ProtoUtil
+import coop.rchain.casper.util.rholang.RuntimeManager
 import coop.rchain.crypto.hash.Blake2b512Random
+import coop.rchain.crypto.signatures.Ed25519
 import coop.rchain.metrics
 import coop.rchain.metrics.Metrics
 import coop.rchain.models.Par
 import coop.rchain.rholang.build.CompiledRholangSource
-import coop.rchain.rholang.interpreter.{accounting, ParBuilder, Runtime, TestRuntime}
 import coop.rchain.rholang.interpreter.Runtime.SystemProcess
 import coop.rchain.rholang.interpreter.accounting.Cost
+import coop.rchain.rholang.interpreter.{accounting, ParBuilder, Runtime, TestRuntime}
 import coop.rchain.shared.Log
 import cats.implicits._
 import coop.rchain.casper.util.rholang.RuntimeManager
@@ -62,6 +67,32 @@ object TestUtil {
       _ <- runtime.shortLeashParams.updateParams(old => old.copy(userId = Par()))
     } yield (runtime)
 
+  def defaultGenesisSetup[F[_]: Concurrent](runtimeManager: RuntimeManager[F]): F[BlockMessage] = {
+    val (_, validators) = (1 to 4).map(_ => Ed25519.newKeyPair).unzip
+    val bonds           = createBonds(validators)
+    val posValidators   = bonds.map(bond => Validator(bond._1, bond._2)).toSeq
+    val ethAddress      = "0x041e1eec23d118f0c4ffc814d4f415ac3ef3dcff"
+    val initBalance     = 37
+    val wallet          = PreWallet(ethAddress, initBalance)
+    //FIXME / TODO: do in Genesis:
+    // - bond deploys for each validator (from their code in Validator)
+    // - create vaults for each validator
+    // - transfers to each validator's vault based on their bonds
+    // - a genesis vault based on the passed public key and initial rev supply
+    //   (introduce new fields to Genesis case class)
+    // - "PoS and testRev" deploys
+    Genesis.createGenesisBlock(
+      runtimeManager,
+      Genesis(
+        "RhoSpec-shard",
+        1,
+        wallet :: Nil,
+        ProofOfStake(0, Long.MaxValue, bonds.map(Validator.tupled).toSeq),
+        false
+      )
+    )
+  }
+
   private def evalDeploy[F[_]: Sync](
       deploy: DeployData,
       runtime: Runtime[F]
@@ -89,5 +120,4 @@ object TestUtil {
       _ <- runtime.reducer.inj(term)
       _ <- runtime.reducer.phlo
     } yield ()
-
 }
